[Bug]: Failed to unsuspend deployment scaled to 0 #90
Bugfix for issue #90

When a deployment or statefulset is scaled to 0, the originalReplicas
annotation will not be set on namespace suspend so we need to handle
this annotation not being present on unsuspend.  Also need to clear out
the originalReplicas annotation on unsuspend so that the replicas can be
set to 0 again.